refactor: split resources into multiple files
